vending-pair-system/backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
pyjwt = "^2.8.0"
python-multipart = "^0.0.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
